+++ main/managers/settings_manager.c
@@ -109,7 +109,8 @@ void settings_set_defaults(FSettings *settings) {
   settings->rgb_mode = RGB_MODE_NORMAL;
   settings->channel_delay = 1.0f;
   settings->broadcast_speed = 5;
-  settings->menu_theme = 0;
+  // default to the 'Bright' palette (index 3)
+  settings->menu_theme = 3;
   strcpy(settings->ap_ssid, "GhostNet");
   strcpy(settings->ap_password, "GhostNet");
   settings->rgb_speed = 15;